Remove IntermediateModelManifest from post processor info
