+++ src/Microsoft.DocAsCode.Build.Engine/Incrementals/ProcessorInfo.cs
@@ -4,6 +4,9 @@
 namespace Microsoft.DocAsCode.Build.Engine.Incrementals
 {
     using System.Collections.Generic;
+    using System.IO;
+
+    using Newtonsoft.Json;
 
     public class ProcessorInfo : ProcessorInfoBase
     {
@@ -11,5 +14,31 @@ namespace Microsoft.DocAsCode.Build.Engine.Incrementals
         /// The information for steps.
         /// </summary>
         public List<ProcessorStepInfo> Steps { get; } = new List<ProcessorStepInfo>();
+        /// <summary>
+        /// The file link for the BuildModel manifest file(type is <see cref="ModelManifest"/>).
+        /// </summary>
+        public string IntermediateModelManifestFile { get; set; }
+        /// <summary>
+        /// Deserialized build intermediate model manifest.
+        /// </summary>
+        [JsonIgnore]
+        public ModelManifest IntermediateModelManifest { get; private set; } = new ModelManifest();
+
+        internal void Load(string baseDir)
+        {
+            if (IntermediateModelManifestFile != null)
+            {
+                IntermediateModelManifest = IncrementalUtility.LoadIntermediateFile<ModelManifest>(Path.Combine(baseDir, IntermediateModelManifestFile));
+            }
+        }
+
+        internal void Save(string baseDir)
+        {
+            if (IntermediateModelManifestFile == null)
+            {
+                IntermediateModelManifestFile = IncrementalUtility.CreateRandomFileName(baseDir);
+            }
+            IncrementalUtility.SaveIntermediateFile(Path.Combine(baseDir, IntermediateModelManifestFile), IntermediateModelManifest);
+        }
     }
 }
