+++ src/Microsoft.DocAsCode.Build.Engine/Incrementals/ProcessorInfoBase.cs
@@ -3,10 +3,6 @@
 
 namespace Microsoft.DocAsCode.Build.Engine.Incrementals
 {
-    using System.IO;
-
-    using Newtonsoft.Json;
-
     public class ProcessorInfoBase
     {
         /// <summary>
@@ -17,31 +13,5 @@ namespace Microsoft.DocAsCode.Build.Engine.Incrementals
         /// The context hash for incremental.
         /// </summary>
         public string IncrementalContextHash { get; set; }
-        /// <summary>
-        /// The file link for the BuildModel manifest file(type is <see cref="ModelManifest"/>).
-        /// </summary>
-        public string IntermediateModelManifestFile { get; set; }
-        /// <summary>
-        /// Deserialized build intermediate model manifest.
-        /// </summary>
-        [JsonIgnore]
-        public ModelManifest IntermediateModelManifest { get; private set; } = new ModelManifest();
-
-        internal void Load(string baseDir)
-        {
-            if (IntermediateModelManifestFile != null)
-            {
-                IntermediateModelManifest = IncrementalUtility.LoadIntermediateFile<ModelManifest>(Path.Combine(baseDir, IntermediateModelManifestFile));
-            }
-        }
-
-        internal void Save(string baseDir)
-        {
-            if (IntermediateModelManifestFile == null)
-            {
-                IntermediateModelManifestFile = IncrementalUtility.CreateRandomFileName(baseDir);
-            }
-            IncrementalUtility.SaveIntermediateFile(Path.Combine(baseDir, IntermediateModelManifestFile), IntermediateModelManifest);
-        }
     }
 }
